requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.3.0
cnocr==2.3.2.2
cnstd==1.2.6.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
doclayout_yolo==0.0.4
filelock==3.18.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.58.4
frozenlist==1.8.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-genai==1.46.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.15.2
MarkupSafe==3.0.2
matplotlib==3.10.3
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.3.1
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
optimum==2.0.0
optimum-onnx==0.0.3
packaging==25.0
pandas==2.3.0
pi_heif==1.1.1
pillow==11.2.1
pillow-avif-plugin==1.5.2
pix2text==1.1.4
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
pyparsing==3.2.3
pyspellchecker==0.8.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.2
rapidocr==3.4.2
regex==2025.10.23
requests==2.32.4
requests-toolbelt==1.0.0
roboflow==1.2.11
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.0
seaborn==0.13.2
sentry-sdk==2.42.1
setuptools==80.9.0
shapely==2.1.2
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stringzilla==4.2.1
sympy==1.14.0
tenacity==9.1.2
thop==0.1.1.post2209072238
tokenizers==0.21.4
torch==2.7.1
torchmetrics==1.8.2
torchvision==0.22.1
tqdm==4.67.1
transformers==4.55.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.159
ultralytics-thop==2.0.14
Unidecode==1.4.0
urllib3==2.5.0
wandb==0.22.2
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
